+++ modulo 3/projeto-astromatch/src/App.js
@@ -7,7 +7,7 @@ import { BASE_URL } from "./constants/url";
 const App = () => {
 
   const [profile, setProfile] = useState({})
-  const [match, setMatch] = useState("")
+  
   
 useEffect(() => {
   getPerfil()
@@ -34,22 +34,15 @@ const getPerfil = () => {
   })
 }
 
-const getMetch = () => {
- axios.get(`${BASE_URL}matches`)
-  .then((response) => {
-    console.log(response.data)
-  })
-  .catch((error) => {
-    console.log(error.response)
-  })
-}
+
 
 
   return (
     <div>
       <Header/>
       <button onClick={clearMatch}>Apagar Match</button>
-      <button onClick={getMetch}>Dar Match</button>
+      <button >Dar Match</button>
+      <getPerfil/>
     </div>
   );
 }
